// src/interfaces/algolia.ts
export interface Hit {
  points: number;
  title: string;
  description: string;
  taster_name: string;
  taster_twitter_handle: string;
  price: number;
  variety: string;
  designation: string;
  region_1: string;
  region_2: string;
  province: string;
  country: string;
  winery: string;
  objectID: string;
}